Feature/endre oppdatere underhold (#602)
* Fikse oppdatering av underhold

* Fikse sletting av notat i fbm sletting av andre barn

+++ src/main/kotlin/no/nav/bidrag/behandling/service/UnderholdService.kt
@@ -66,14 +66,21 @@ class UnderholdService(
     ): Underholdskostnad {
         gjelderBarn.validere()
         gjelderBarn.personident?.let { personidentBarn ->
-            val rolle = behandling.søknadsbarn.find { it.ident == personidentBarn.verdi }
+            val rolleSøknadsbarn = behandling.søknadsbarn.find { it.ident == personidentBarn.verdi }
             val eksisterendePerson = personRepository.findFirstByIdent(personidentBarn.verdi)
             if (eksisterendePerson == null) {
-                val roller = rolle?.let { mutableSetOf(it) } ?: mutableSetOf()
-                val person = personRepository.save(Person(ident = personidentBarn.verdi, rolle = roller))
+                val person =
+                    personRepository.save(
+                        Person(
+                            ident = personidentBarn.verdi,
+                            rolle = rolleSøknadsbarn?.let { mutableSetOf(it) } ?: mutableSetOf(),
+                        ),
+                    )
                 person.rolle.forEach { it.person = person }
                 return lagreUnderholdskostnad(behandling, person)
             } else {
+                rolleSøknadsbarn?.let { eksisterendePerson.rolle.add(it) }
+                rolleSøknadsbarn?.person = eksisterendePerson
                 return lagreUnderholdskostnad(behandling, eksisterendePerson)
             }
         } ?: run {
@@ -260,7 +267,11 @@ class UnderholdService(
         underholdskostnad: Underholdskostnad,
     ): UnderholdDto? {
         behandling.underholdskostnader.remove(underholdskostnad)
-        personRepository.deleteById(underholdskostnad.person.id!!)
+        underholdskostnad.person.underholdskostnad.remove(underholdskostnad)
+        if (underholdskostnad.person.underholdskostnad.isEmpty() && underholdskostnad.person.rolle.isEmpty()) {
+            personRepository.deleteById(underholdskostnad.person.id!!)
+            notatService.sletteNotat(behandling, Notattype.UNDERHOLDSKOSTNAD, behandling.bidragsmottaker!!)
+        }
         underholdskostnadRepository.deleteById(underholdskostnad.id!!)
         return null
     }
